# backend/docs/swagger/swagger.yaml
basePath: /
definitions:
...
  dtos.TuyaSyncDeviceDTO:
    properties:
      create_time:
        type: integer
      id:
        type: string
      online:
        type: boolean
      remote_id:
        type: string
      update_time:
        type: integer
    type: object
...
  sensio_domain_common_dtos.StandardResponse:
    properties:
      data: {}
      details:
        description: |-
          Details is only populated for 400 (Bad Request) and 422 (Unprocessable Entity) errors.
          For all other status codes, including 500, this field is nil/omitted.
      message:
        example: Success
        type: string
      status:
        example: true
        type: boolean
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for Sensio App. <br> For full documentation,
    visit <a href="/docs">/docs</a>.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sensio API
  version: "1.0"
paths:
...
  /api/devices/{id}/statuses:
    get:
      consumes:
      - application/json
      description: GetDeviceStatusesByDeviceID
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: GetDeviceStatusesByDeviceID
      tags:
      - 09. Terminals
...
  /api/devices/terminal/{terminal_id}:
    get:
      consumes:
      - application/json
      description: GetDevicesByTerminalID
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: GetDevicesByTerminalID
      tags:
      - 09. Terminals
...
  /api/terminal/{id}:
    delete:
      consumes:
      - application/json
      description: DeleteTerminal
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: DeleteTerminal
      tags:
      - 09. Terminals
    get:
      consumes:
      - application/json
      description: GetTerminalByID
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: GetTerminalByID
      tags:
      - 09. Terminals
    put:
      consumes:
      - application/json
      description: UpdateTerminal
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: UpdateTerminal
      tags:
      - 09. Terminals
...
  /api/tuya/devices/sync:
    get:
      consumes:
      - application/json
      description: Fetches real-time status from Tuya. Does NOT update local DB. Returns
        fresh device list.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/sensio_domain_common_dtos.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dtos.TuyaSyncDeviceDTO'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/sensio_domain_common_dtos.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sensio_domain_common_dtos.StandardResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Sync Terminal Device Status
      tags:
      - 02. Tuya
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-KEY
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...